head pointer can pass tail pointer
The latest commit modified head pointer so it does not wrap around
anymore. This is problematic, since head can pass tail pointer and
since in 'Push' head is being compared to tail, this can create
issues when the buffer is expanded and then values are popped from
the queue. The problem is that then the tail never gets equal to
head, and both tail and head wrap around, which can result in
overwrites or duplicated messages popped form the queue.